+++ admin/resources/projects/projects.resource.ts
@@ -7,7 +7,7 @@ import {
 } from "adminjs";
 import { dataSource } from "../../datasource.js";
 import { COMMON_RESOURCE_LIST_PROPERTIES } from "../common/common.resources.js";
-import { Project } from "@shared/entities/users/projects.entity.js";
+import { Project } from "@shared/entities/projects.entity.js";
 import { Country } from "@shared/entities/country.entity.js";
 
 export const ProjectsResource: ResourceWithOptions = {
