+++ api/src/modules/projects/projects.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Project } from '@shared/entities/users/projects.entity';
+import { Project } from '@shared/entities/projects.entity';
 import { ProjectsController } from './projects.controller';
 import { ProjectsService } from './projects.service';
