+++ shared/lib/db-entities.ts
@@ -24,7 +24,7 @@ import { EmissionFactors } from "@shared/entities/emission-factors.entity";
 import { LongTermProjectOperating } from "@shared/entities/long-term-project-operating.entity";
 import { MRV } from "@shared/entities/mrv.entity";
 import { SequestrationRate } from "@shared/entities/sequestration-rate.entity";
-import { Project } from "@shared/entities/users/projects.entity";
+import { Project } from "@shared/entities/projects.entity";
 
 export const COMMON_DATABASE_ENTITIES = [
   User,
